cleanup; php 7.2 - 8.2 ci setup; fix compatibility
